feat: h264 parser check keyframe

+++ h264/include/h264-hrd.h
@@ -5,7 +5,7 @@
 
 struct h264_hrd_t
 {
-	uint32_t cpb_cnt_minus1;
+	uint8_t cpb_cnt_minus1; // [0, 32]
 	uint32_t bit_rate_value_minus1[32];
 	uint32_t cpb_size_value_minus1[32];
 	uint32_t cbr_flag[32];
